requirements.txt
alembic==1.14.0
black==24.10.0
celery==5.4.0
docx==0.2.4
fastapi==0.115.5
flower==2.0.1
pydantic==2.10.1
pydantic_core==2.27.1
python-docx==1.1.2
python-dotenv==1.0.1
redis==5.2.0
requests==2.32.3
SQLAlchemy==2.0.36
sqlalchemy-stubs==0.4
httpx==0.27.2
typing_extensions==4.12.2
uvicorn==0.32.1
websockets==14.1
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==5.0.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
docx==0.2.4
fastapi==0.115.5
flower==2.0.1
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
humanize==4.11.0
idna==3.10
kombu==5.4.2
lxml==5.3.0
Mako==1.3.6
MarkupSafe==3.0.2
mypy==1.13.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.0.0
platformdirs==4.3.6
prometheus_client==0.21.0
prompt_toolkit==3.0.48
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-magic==0.4.27
pytz==2024.2
redis==5.2.0
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlalchemy-stubs==0.4
starlette==0.41.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.1
vine==5.1.0
wcwidth==0.2.13
websockets==14.1
